refactor: Refactor auth-related things to auth.controller
