+++ server/config/routes.js
@@ -1,15 +1,8 @@
 var controller = require('./controller');
-var auth = require('./auth');
+var auth = require('./auth.controller');
 
 module.exports = function(express, passport) {
   var routes = express.Router();
-
-  routes.get('/auth/facebook', passport.authenticate('facebook'));
-
-  routes.get('/auth/facebook/callback', passport.authenticate('facebook', {
-    successRedirect: '/',
-    failureRedirect: '/fail'
-  }));
 
   routes.get('/fail', function(req, res){
     res.end('"ok"');
