+++ server/server.js
@@ -3,7 +3,7 @@ const express = require('express');
 const morgan = require('morgan');
 const session = require('express-session');
 
-const auth = require('./config/auth');
+const auth = require('./config/auth.controller');
 const routes = require('./config/routes');
 
 let app = express();
@@ -15,8 +15,7 @@ app.use(session({
   resave: false,
   saveUninitialized: false
 }));
-app.use(auth.passport.initialize());
-app.use(auth.passport.session());
+auth.loadController(app);
 app.use(routes(express, auth.passport));
 app.use(express.static(__dirname + '/../client/public'));
